+++ src/provider/fileProvider.ts
@@ -3,7 +3,7 @@ import * as path from "path";
 import * as vscode from "vscode";
 
 import { SettingsProvider } from "./settingsProvider";
-import { JsonInterface, pathExists, readJsonFile } from "./utils";
+import { JsonInterface, pathExists, readJsonFile } from "../utils";
 
 export abstract class FileProvider {
   public templatePath: string;
@@ -24,7 +24,7 @@ export abstract class FileProvider {
     const deletePattern = `${this.vscodeDirectory}/**`;
 
     const extDirectory = path.dirname(__dirname);
-    const templateDirectory = path.join(extDirectory, "src", "templates");
+    const templateDirectory = path.join(extDirectory, "src", "_templates");
     this.templatePath = path.join(templateDirectory, templateFileName);
 
     this.fileWatcherOnDelete = vscode.workspace.createFileSystemWatcher(
@@ -74,7 +74,7 @@ export abstract class FileProvider {
     this.writeFileData(this.outputPath, this.outputPath);
   }
 
-   // @ts-ignore
+  // @ts-ignore
   public writeFileData(inputFilePath: string, outFilePath: string) {
     throw new Error("You have to implement the method doSomething!");
   }